feature(update) : refactor package managers
