+++ lib/package-managers/npm.package-manager.ts
@@ -11,39 +11,6 @@ export class NpmPackageManager extends AbstractPackageManager {
     super(RunnerFactory.create(Runner.NPM, logger), logger);
   }
 
-  public async install(directory: string) {
-    const spinner = ora({
-      spinner: {
-        "interval": 120,
-        "frames": [
-          "▹▹▹▹▹",
-          "▸▹▹▹▹",
-          "▹▸▹▹▹",
-          "▹▹▸▹▹",
-          "▹▹▹▸▹",
-          "▹▹▹▹▸"
-        ]
-      },
-      text: messages.PACKAGE_MANAGER_INSTALLATION_IN_PROGRESS
-    });
-    spinner.start();
-
-    try {
-      await super.install(directory)
-      spinner.succeed();
-      this.logger.info();
-      this.logger.info(messages.PACKAGE_MANAGER_INSTALLATION_SUCCEED(directory));
-      this.logger.info(messages.GET_STARTED_INFORMATION);
-      this.logger.info();
-      this.logger.info(chalk.gray(messages.CHANGE_DIR_COMMAND(directory)));
-      this.logger.info(chalk.gray(messages.START_COMMAND));
-      this.logger.info();
-    } catch {
-      spinner.fail();
-      this.logger.error(chalk.red(messages.PACKAGE_MANAGER_INSTALLATION_FAILED));
-    }
-  }
-
   public get name() {
     return PackageManager.NPM.toUpperCase();
   }
