+++ lib/package-managers/yarn.package-manager.ts
@@ -11,39 +11,6 @@ export class YarnPackageManager extends AbstractPackageManager {
     super(RunnerFactory.create(Runner.YARN, logger), logger);
   }
 
-  public async install(directory: string) {
-    const spinner = ora({
-      spinner: {
-        "interval": 120,
-        "frames": [
-          "▹▹▹▹▹",
-          "▸▹▹▹▹",
-          "▹▸▹▹▹",
-          "▹▹▸▹▹",
-          "▹▹▹▸▹",
-          "▹▹▹▹▸"
-        ]
-      },
-      text: messages.PACKAGE_MANAGER_INSTALLATION_IN_PROGRESS
-    });
-    spinner.start();
-    super.install(directory)
-      .then(() => {
-        spinner.succeed();
-        this.logger.info();
-        this.logger.info(messages.PACKAGE_MANAGER_INSTALLATION_SUCCEED(directory));
-        this.logger.info(messages.GET_STARTED_INFORMATION);
-        this.logger.info();
-        this.logger.info(chalk.gray(messages.CHANGE_DIR_COMMAND(directory)));
-        this.logger.info(chalk.gray(messages.START_COMMAND));
-        this.logger.info();
-      })
-      .catch(() => {
-        const message = messages.PACKAGE_MANAGER_INSTALLATION_FAILED;
-        this.logger.error(chalk.red(message));
-      });
-  }
-
   public get name() {
     return PackageManager.YARN.toUpperCase();
   }
